Update README.md
What was noticed by nader junaid syrian pilot for a
long time age many notices the most obvious one
when he said The path of the moon's shadow during
the last solar eclipse was from the west to the east,
and this can not be if the earth revolves around its
self
I have noticed that eddington When he proved the
theory of relativity in 1919, where compared two
images of the sky the first during the eclipse of the
sun in the day the second got during the night after
that day and found Gap between the positions of the
stars in the two images. he concluded the curvature
of light near the gravitational field
No one noticed since that date, he could not see the
same scene in the day and in the night that followed
if the earth revolves around its self . Where we must
see the other half of the sky . This may be we can
consider it contrary to the theory of Kepler clearly,
but the problem that we face is how to find a model
can explain the universe
I found during research for a master's degree, which
was about (gravitational lensing and creating a
cosmic ruler) redshifts corresponding wedth
Einsteins ring for cosmic objects

`

the most impressive thing that I have noticed
during my study to get the master degree in physics
is the fantastic graph that I made between the
redshift and the width of the dark mater
(by means of Einstein ring
a small change in redshift taken the same along
the scale corresponding increasing changes
exponentially
the width of dark matter this could not be
understood unless we assumed that the
conservation law of mass and momentum in every
gravitational field is independent and from our view
in the earth; the things will be smaller if we go far
away from our planet and the distances will be
reduced which means the Kg will be lesser and
lesser and the meter will reach to be a cent
all this means that all our understanding about the
universe and its distances is wrongWhat was noticed by nader junaid syrian pilot for a\nlong time age many notices the most obvious one\nwhen he said The path of the moon\'s shadow during\nthe last solar eclipse was from the west to the east,\nand this can not be if the earth revolves around its\nself\nI have noticed that eddington When he proved the\ntheory of relativity in 1919, where compared two\nimages of the sky the first during the eclipse of the\nsun in the day the second got during the night after\nthat day and found Gap between the positions of the\nstars in the two images. he concluded the curvature\nof light near the gravitational field\nNo one noticed since that date, he could not see the\nsame scene in the day and in the night that followed\nif the earth revolves around its self . Where we must\nsee the other half of the sky . This may be we can\nconsider it contrary to the theory of Kepler clearly,\nbut the problem that we face is how to find a model\ncan explain the universe\nI found during research for a master\'s degree, which\nwas about (gravitational lensing and creating a\ncosmic ruler) redshifts corresponding wedth\nEinsteins ring for cosmic objects\n`\nthe most impressive thing that I have noticed\nduring my study to get the master degree in physics\nis the fantastic graph that I made between the\nredshift and the width of the dark mater\n(by means of Einstein ring\na small change in redshift taken the same along\nthe scale corresponding increasing changes\nexponentially\nthe width of dark matter this could not be\nunderstood unless we assumed that the\nconservation law of mass and momentum in every\ngravitational field is independent and from our view\nin the earth; the things will be smaller if we go far\naway from our planet and the distances will be\nreduced which means the Kg will be lesser and\nlesser and the meter will reach to be a cent\nall this means that all our understanding about the\nuniverse and its distances is wrong